requirements.txt
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
Django==6.0.1
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
lxml==6.0.2
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.3
platformdirs==4.5.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.7
python-docx==1.2.0
python-dotenv==1.2.1
pytokens==0.4.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
